[misc][case] separated card dropdown component into another folder
