+++ src/components/CardDropdown/styles.scss
@@ -1,0 +1,12 @@
+.menu-icon {
+  .ant-dropdown-menu-item-icon {
+    max-width: 12px;
+  }
+}
+
+.ant-card-extra {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+}
